Represented Analysis text in component
