--- PennyPincherUI/src/app/dashboard/analysis/analysis.component.html
@@ -1,47 +1,0 @@
-<p>analysis works!</p>
-
-<h1>Financial Summary</h1>
-
-<div>
-  <h2>Status:</h2>
-  <table>
-    <thead>
-      <tr>
-        <th>GrossIncome</th>
-        <th>NetIncome</th>
-        <th>Liabilities</th>
-        <th>NetIncomeRatio</th>
-        <th>MostCostlyName</th>
-        <th>MostCostlyAmount</th>
-        <th>PercentOfEarningsGoingToMostCostlyAmount</th>
-      </tr>
-    </thead>
-    <tbody>
-      <tr>
-        <td>
-          {{ statuses().grossIncome }}
-        </td>
-        <td>
-          {{ statuses().netIncome }}
-        </td>
-        <td>
-          {{ statuses().liabilities }}
-        </td>
-        <td>
-          {{ statuses().netIncomeRatio }}
-        </td>
-        <td>
-          {{ statuses().mostCostlyName }}
-        </td>
-        <td>
-          {{ statuses().mostCostlyAmount }}
-        </td>
-        <td>
-          {{ statuses().percentOfEarningsGoingToMostCostlyAmount }}
-        </td>
-      </tr>
-    </tbody>
-  </table>
-  <h3>this is some data: {{ this.statuses().grossIncome }}</h3>
-  <button (click)="printStatus()">Get Status</button>
-</div>
