--- PennyPincherUI/src/app/dashboard/analysis/analysis.component.spec.ts
@@ -1,30 +1,0 @@
-import { type ComponentFixture, TestBed } from "@angular/core/testing";
-
-import { AnalysisComponent } from "./analysis.component";
-import {
-	HttpTestingController,
-	provideHttpClientTesting,
-} from "@angular/common/http/testing";
-import { provideHttpClient } from "@angular/common/http";
-
-describe("AnalysisComponent", () => {
-	let component: AnalysisComponent;
-	let fixture: ComponentFixture<AnalysisComponent>;
-	let httpTesting: HttpTestingController;
-
-	beforeEach(async () => {
-		await TestBed.configureTestingModule({
-			imports: [AnalysisComponent],
-			providers: [provideHttpClient(), provideHttpClientTesting()],
-		}).compileComponents();
-
-		httpTesting = TestBed.inject(HttpTestingController);
-		fixture = TestBed.createComponent(AnalysisComponent);
-		component = fixture.componentInstance;
-		fixture.detectChanges();
-	});
-
-	it("should create", () => {
-		expect(component).toBeTruthy();
-	});
-});
