+++ PennyPincherUI/src/app/expenses/expense-tables/expense-table.component.html
@@ -22,7 +22,7 @@
               <th>Name</th>
               <th>Category</th>
               <th>Price</th>
-              <th>Delete?</th>
+              <th>Delete</th>
             </tr>
           </thead>
           <tbody>
